# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  title: Api
  version: 0.1.0
  description: Choice Credit and Rental Solutions API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /clients/me:
    get:
      operationId: getMyProfile
      tags: [client]
      summary: Get current client profile
      responses:
        "200":
          description: Client profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
        "401":
          description: Unauthorized
    put:
      operationId: updateMyProfile
      tags: [client]
      summary: Update client profile
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ProfileUpdate"
      responses:
        "200":
          description: Updated profile
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ClientProfile"
  /clients/me/case:
    get:
      operationId: getMyCase
      tags: [client]
      summary: Get client's active case
      responses:
        "200":
          description: Active case
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Case"
        "404":
          description: No active case
  /clients/me/documents:
    get:
      operationId: listMyDocuments
      tags: [client]
      summary: List client documents
      responses:
        "200":
          description: Document list
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Document"
    post:
      operationId: uploadDocument
      tags: [client]
      summary: Upload a document
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/DocumentInput"
      responses:
        "201":
          description: Document uploaded
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Document"
...
  /clients/me/plan:
    get:
      operationId: getMyPlan
      tags: [client]
      summary: Get personalized approval roadmap
      responses:
        "200":
          description: Approval plan
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ApprovalPlan"
        "404":
          description: No plan yet
  /clients/me/messages:
    get:
      operationId: listMyMessages
      tags: [client]
      summary: Get client messages
      responses:
        "200":
          description: Messages
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Message"
    post:
      operationId: sendMessage
      tags: [client]
      summary: Send a message to advisor
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/MessageInput"
      responses:
        "201":
          description: Message sent
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Message"
...
  /clients/me/payment-selection:
    post:
      operationId: selectPaymentMethod
      tags: [client]
      summary: Select preferred payment method for invoice
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/PaymentSelection"
      responses:
        "201":
          description: Payment method selected
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/PaymentSelectionResult"
...
components:
  schemas:
...
    ClientProfile:
      type: object
      required: [id, email, fullName, role, createdAt]
      properties:
        id:
          type: string
        email:
          type: string
        fullName:
          type: string
        phone:
          type: ["string", "null"]
        role:
          type: string
          enum: [client, admin]
        avatarUrl:
          type: ["string", "null"]
        createdAt:
          type: string
    ProfileUpdate:
      type: object
      properties:
        fullName:
          type: string
        phone:
          type: ["string", "null"]
    Case:
      type: object
      required: [id, clientId, status, packageName, createdAt]
      properties:
        id:
          type: string
        clientId:
          type: string
        packageName:
          type: string
        status:
          type: string
          enum: [intake, in_progress, review, approved, completed]
        advisorName:
          type: ["string", "null"]
        notes:
          type: ["string", "null"]
        createdAt:
          type: string
        updatedAt:
          type: string
    Document:
      type: object
      required: [id, clientId, name, fileUrl, fileType, status, uploadedAt]
      properties:
        id:
          type: string
        clientId:
          type: string
        caseId:
          type: ["string", "null"]
        name:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        status:
          type: string
          enum: [pending, reviewing, approved, rejected]
        advisorNotes:
          type: ["string", "null"]
        uploadedAt:
          type: string
    DocumentInput:
      type: object
      required: [name, fileUrl, fileType]
      properties:
        name:
          type: string
        fileUrl:
          type: string
        fileType:
          type: string
        caseId:
          type: ["string", "null"]
...
    ApprovalPlan:
      type: object
      required: [id, caseId, title, steps, createdAt]
      properties:
        id:
          type: string
        caseId:
          type: string
        title:
          type: string
        steps:
          type: array
          items:
            $ref: "#/components/schemas/PlanStep"
        createdAt:
          type: string
        updatedAt:
          type: string
    PlanStep:
      type: object
      required: [id, order, title, status]
      properties:
        id:
          type: string
        order:
          type: integer
        title:
          type: string
        description:
          type: ["string", "null"]
        status:
          type: string
          enum: [pending, in_progress, completed]
        dueDate:
          type: ["string", "null"]
...
    Message:
      type: object
      required: [id, senderId, senderName, content, createdAt]
      properties:
        id:
          type: string
        caseId:
          type: ["string", "null"]
        clientId:
          type: ["string", "null"]
        senderId:
          type: string
        senderName:
          type: string
        senderRole:
          type: string
          enum: [client, admin]
        content:
          type: string
        createdAt:
          type: string
        readAt:
          type: ["string", "null"]
    MessageInput:
      type: object
      required: [content]
      properties:
        content:
          type: string
        caseId:
          type: ["string", "null"]
...
    PaymentSelection:
      type: object
      required: [invoiceId, method]
      properties:
        invoiceId:
          type: string
        method:
          type: string
          enum: [cashapp, paypal, zelle, applepay, googlepay, venmo]
    PaymentSelectionResult:
      type: object
      required: [invoiceId, method, message]
      properties:
        invoiceId:
          type: string
        method:
          type: string
        message:
          type: string
...