# lib/api-zod/src/generated/types/paymentSelection.ts
/**
 * Generated by orval v8.18.0 🍺
 * Do not edit manually.
 * Api
 * Choice Credit and Rental Solutions API
 * OpenAPI spec version: 0.1.0
 */
import type { PaymentSelectionMethod } from './paymentSelectionMethod';

export interface PaymentSelection {
  invoiceId: string;
  method: PaymentSelectionMethod;
}